refactor(electron): split SuccessView into wt-* subcomponents [WTEL-9965](https://webitel.atlassian.net/browse/WTEL-9965)
- move SuccessView into SuccessView/ dir matching CallView pattern
- replace dane checkmark bg-image with wt-icon (done, success)
- replace Close button with wt-button (warn)
- flex-column layout, drop DOM id, inline styles, hand-rolled CSS
- dedup timer cleanup into clearTimer helper
